fix(ivy): ensure inherited style/class bindings work in both parent/child surroundings
Prior to this fix, styling bindings inherited from a parent component would conflict
with any bindings present in the child class component. The reason for this was because
the same directive instance (which is referenced within the AOT code for styling host
bindings) was used for both the parent and child host bindings function. This means that
there wasn't anything to differentiate between the two sets of bindings.

This patch makes sure that the Angular styling algorithm understands when style bindings
are being evaluated within the parent component's host binding function and when it runs
within the child area.

Jira Issue: FW-1132