fix(ivy): ensure styling instructions are automatically flushed once CD exits per element
Prior to this fix, all styling instructions would only be applied if a
`elementStylingApply` or `elementHostStylingApply` instruction is
executed. To salvage instruction space, Angular will now not emit the
instruction from the compiler and will know internally to render
styling depending if any styles or classes have already been set.