feat(ivy): introduce `select()` to incrementally invoke directive hostBindings and hooks
Prior to this fix, all directive/component host bindings update code was
run AFTER the template function for the parent component was run. This
means that the associated host bindings for an element that has a
directive and/or component attached to it will not run inline with the
bindings on that element before the next element is processed. This
patch ensures that the all host bindings for an element are processed
before the next element is visited.